Add an example of if there's a single identity provider we automatically redirect the user to them. If there's multiple identity providers then we need to redirect to a login page to get the username (to figure out which identity provider to send them to)
TODO - does forcing this kind of redirect make since for an API? e.g. non-browser programmatic access wouldn't be able to accept the redirects here and should probably send a 401 instead of a redirect. This can probably be handled by considering the context of the request and using a different  AuthenticationEntryPoint that then sends back a 401 instead of a redirect.